BASE

+++ veterinaryWebServices/src/main/java/com/team12/veterinaryWebServices/viewmodel/profileVM.java
@@ -1,9 +1,8 @@
 package com.team12.veterinaryWebServices.viewmodel;
 
 import com.team12.veterinaryWebServices.model.profile;
-import com.team12.veterinaryWebServices.model.role;
 
-public record profileVM(Long profileID, String roleNAME, String profileNAME, String profileEMAIL, boolean GENDER, int AGE, String PHONE) {
+public record profileVM(Long profileID, String roleNAME, String profileNAME, String profileEMAIL, boolean GENDER, Long AGE, String PHONE) {
     public static profileVM from (profile p){
 
         return new profileVM(
